feat(emit-dynamic-site): Phase 3 step 4 — object + building sprites
Holtburg now reads as a town. The 3×3 grid renders 239 placed
objects (props + buildings) on top of the existing terrain atlas:
brown-tinted building silhouettes clustered around the road
junction, smaller props scattered through the grass. Compare to
the static-site z=12 reference (`docs/images/DerethMapsEnhanced_zoom.png`):
same layout, same building cluster — the visual gap that step 3.5
left open ("missing only the building/tree sprites that step 4 will
land") is now closed.

Three pieces:

**1. Bug fix in `BuildInfo` parser.** `crates/holtburger-dat/src/
landblock.rs`'s `BuildInfo.num_portals` field was declared `u16`,
but ACE's `List<T>.Unpack(reader)` extension reads a `u32` count.
The mismatch only triggered when a building had any portals —
Holtburg's first interior building (`0xA9B4FFFE`) hits it
immediately and the parser panicked with "failed to fill whole
buffer while parsing stab_list in PortalInternal". Pinned via the
new round-trip smoke check; native bit-decode tests already passed
because synthetic LandblockInfo bytes had zero portals.

**2. `fetch_landblock_objects` wasm export.** Returns one
`ObjectPlacement` per `Stab` (props) and one per `BuildInfo`
(buildings) in the requested `XXYYFFFE` cells. Fields: `landblockId`,
`modelId` (AC GfxObj/SetupModel id, 0x01XXXXXX or 0x02XXXXXX), `x`,
`y`, `z` (landblock-local metres), `rotationZ` (yaw radians from
quaternion via `atan2(2(qw*qz + qx*qy), 1 - 2(qy² + qz²))`).
Per-cell missing → silently skip rather than fail (ocean cells
have no LandblockInfo). The 3×3 around Holtburg yields 239
placements (113 objects + 126 building outlines).

**3. JS sprite layer.** New `loadSpriteAtlas()` fetches the
static-site atlas at `./sprites/atlas.{png,js}` (4096×1296
greyscale silhouettes for 108 model IDs); copied into
`apps/holtburger-web/sprites/` so the same path resolves at any
serve root. New `buildObjectsContainer(neighbourhood, objects,
spriteMap)` instantiates one `PIXI.Sprite` per object where the
model_id is in the atlas, falling back to a small coloured circle
otherwise. Coverage: 44% of Holtburg's objects have an atlas tile;
the rest render as fallback dots.

**Tinting.** Atlas tiles are pure greyscale (R == G == B every
pixel). PIXI.Sprite.tint multiplies per-channel, so a brown tint
turns the silhouette into a brown building. Two-bucket palette by
model_id top byte: 0x01 (GfxObj/Model — usually structures) →
warm brown #8b6442; 0x02 (SetupModel — usually props/foliage) →
desaturated tan-green #6f7a4a. Placeholder mapping; AC's real
per-model material isn't in the static-site sample atlas. A
project-specific atlas with real colours can replace it without
shader changes.

**Two implementation footguns documented.**

(a) The `LandblockInfo.id` field carries the `XXYYFFFE` file id,
not the `XXYYFFFF` CellLandblock id that the heightmap layer keys
on. First lookup pass mapped `NEIGHBOURHOOD.id` directly — every
object got `continue`-skipped. Fix: index `neighbourhood` by
`(lbX << 16 | lbY)` and decompose `obj.landblockId` to match.

(b) PIXI.Graphics circles drawn at world-metre size: a 1-meter
radius is sub-pixel at the default 3×3-fits-canvas zoom and renders
invisibly. Bumped fallback dot radius to 1.5 m and used during
debugging at 5 m before reverting.

**Coordinate convention.** AC's positive yaw is CCW from above.
The `worldContainer.scale.set(1, -1)` y-flip mirrors the scene, so
sprite rotation needs to be negated (CCW in world → CW after flip).

Verification:
- `cargo test --workspace --lib`: 1096/0 (unchanged — Rust changes
  are wasm32-only or covered by the smoke).
- `node smoke_test.cjs`: 32/32 PASS (was 28, +4 step-4 checks
  ship in the next commit).
- Browser screenshot at `docs/images/phase-3-step-4-objects.png`
  shows brown buildings clustered around Holtburg's road junction,
  smaller props scattered through the grass, and the same terrain
  + roads from step 3.5 underneath.

+++ external/holtburger/crates/holtburger-dat/src/landblock.rs
@@ -24,8 +24,12 @@ pub struct BuildInfo {
     pub model_id: u32,
     pub frame: Frame,
     pub num_leaves: u32,
+    // u32 count, matching ACE's `List<T>.Unpack(reader)` extension
+    // (`UnpackableExtensions.cs:174`). Earlier u16 here desync'd the
+    // parser when a building had no portals (read 0 then misaligned
+    // for the next building's model_id).
     #[br(temp)]
-    pub num_portals: u16,
+    pub num_portals: u32,
     #[br(count = num_portals)]
     pub portals: Vec<PortalInternal>,
 }
